+++ blueprints/codes/eurocode/en_1993_1_1_2022/chapter_8_ultimate_limit_state/formula_8_13.py
@@ -3,7 +3,7 @@
 import operator
 from collections.abc import Callable
 
-from blueprints.codes.eurocode.en_1993_1_1_2025 import EN_1993_1_1_2025
+from blueprints.codes.eurocode.en_1993_1_1_2022 import EN_1993_1_1_2022
 from blueprints.codes.formula import ComparisonFormula
 from blueprints.codes.latex_formula import LatexFormula, latex_replace_symbols
 from blueprints.type_alias import N
@@ -14,7 +14,7 @@ class Form8Dot13UnityCheckTensileStrength(ComparisonFormula):
     r"""Class representing formula 8.13 for the unity check for tensile strength."""
 
     label = "8.13"
-    source_document = EN_1993_1_1_2025
+    source_document = EN_1993_1_1_2022
 
     def __init__(
         self,
