+++ blueprints/codes/eurocode/en_1993_1_1_2022/chapter_8_ultimate_limit_state/formula_8_14.py
@@ -1,6 +1,6 @@
 """Formula 8.14 from EN 1993-1-1:2025: Chapter 8 - Ultimate Limit State."""
 
-from blueprints.codes.eurocode.en_1993_1_1_2025 import EN_1993_1_1_2025
+from blueprints.codes.eurocode.en_1993_1_1_2022 import EN_1993_1_1_2022
 from blueprints.codes.formula import Formula
 from blueprints.codes.latex_formula import LatexFormula, latex_replace_symbols
 from blueprints.type_alias import DIMENSIONLESS, MM2, MPA, N
@@ -11,7 +11,7 @@ class Form8Dot14DesignPlasticRestistanceGrossCrossSection(Formula):
     r"""Class representing formula 8.14 for the calculation of [$N_{pl,Rd}$]."""
 
     label = "8.14"
-    source_document = EN_1993_1_1_2025
+    source_document = EN_1993_1_1_2022
 
     def __init__(
         self,
