+++ tests/codes/eurocode/en_1993_1_1_2022/chapter_8_ultimate_limit_state/__init__.py
@@ -1,1 +1,1 @@
-"""Module containing all formulas from EN_1993_1_1_2025: Chapter 6 - Ultimate limit state."""
+"""Module containing all formulas from EN_1993_1_1_2022: Chapter 6 - Ultimate limit state."""
